+++ packages/hooks/use-ripple/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @frontplus-ui/use-ripple
 
+## 1.0.2
+
+### Patch Changes
+
+- fix type issues
+
 ## 1.0.1
 
 ### Patch Changes